+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve handling of cast expressions in `MemoryManagementRule`.
 - Change `memoryFunctions` parameter to fully-qualified names in `MemoryManagementRule`.
 - Remove `MemoryManagementRule` from the Sonar Way quality profile.
+- Improve handling of constructor calls in `RedundantCastRule`.
 
 ## [0.28.0] - 2021-02-02